Move tasks into the corresponding directories
Because the tasks are derivative work, the same license shall apply
as for the original tasks. However we do not know the license in
every case, so the most sensible thing to do is to just put the tasks
next to the originals. This also ensures that they are always run
next to the original tasks and one can compare the results more easily.