First draft of runtime error property
We discussed to add the runtime error reachability as a porperty
to the SV-COMP java track. I propose that the property is
described as CHECK( init(Main.main()), LTL(not (F RuntimeException))
in an LTL formula to express that a runtime error should not be reachable.

This is a proposal and I cannot gurantee that all true/false
verdicts are corrected. I obtained them by using random testing
on the existing tasks.

StackOverflow errors are not considered as runtime errors, as they are JVM specific
and depend on the configuration of the JVM. It is not a property of the program, IMHO.