Fix data race in 43_NetBSD_sysmon_power_sliced.c
The problem is that the cv_wait emulation accesses its variable
without proper locking, i.e., this is an unguarded read.
Any guarded write by another thread will therefore produce a data
race. To reproduce this, you can limit the task to two spawned
threads and the following case:

case 3: sysmonread_power(); break;

When both threads are in the sysmonread_power call, it can easily
happen that one sets COND to 0, while the other reads the value
of COND in the evaluation of assume(c).

This problem is fixed by switching two statements in the cv_wait
preprocessor macro. That way the read will be guarded by the
"lock" set via __VERIFIER_acquire the way this mechanism was
intended. The variable is still able to be changed by the other
thread in between the release and acquire, no problem there.

Note that this bug might be present in other tasks that use the
cv_wait macro as well. Some of them are already flagged as to contain
a data race, though I was not yet able to determine whether this
is actually also a race caused by this macro. If this is the case,
we should probably create a fixed version of those tasks as well.