Make if_etherip harness call ip_deliver with correct protocol element
Two issues in the benchmark task are fixed to do this:

1. The wrong ip_init was called. Instead, the verification harness
   (main method) has to call ip6_init to initialize the data structures for ipv6.

2. ip_deliver parameter 'nxt' should point to the relevant protocol step,
but it was called with value 0.
0 points to element 3 (rip6_input), which is not relevant for us.
We directly call ip6_etherip_input, which contains the error we are looking for.

With this change, the erroneous ip6_etherip_input is called
and the actual intended violation in this verification task
happens: After ip6_etherip_input is finished, it mistakingly does not return
IPPROTO_DONE, but IPPROTO_NONE. This makes it call rip_input
instead of stopping the delivery.

We do not adjust task 'if_etherip-fixed-double-free-invalid-deref1':
it keeps calling ip_init instead of ip6_init, to keep the invalid
dereference bug.