fix type of thread functions in weaver benchmarks
As noted in [1], the declaration of pthread_create in these benchmarks
expects a pointer to a function which takes a void* argument. However,
the functions serving as thread templates in these benchmarks did not
take any argument. This constitutes undefined behaviour.

As a fix, we introduce an unused argument `void* _argptr` for all such
functions.

The change was made with the following command:

  sed -i 's/void\* thread\(.\)()/void* thread\1(void* _argptr)/g' *.wvr.c

[1] https://gitlab.com/sosy-lab/benchmarking/sv-benchmarks/-/issues/1362