Change syntax for inclusion of job rules in GitLab's CI config
The previous syntax declared "rules" as a list of one element,
and this one element included another list of rules via !reference.
This lead to a nested list of rules (actually visible via the
"Full configuration" tab of GitLab's CI editor.

This seemed to work fine so far, but there is #1390,
and maybe this is caused by the nested list of rules,
because inside .branch-rules we have two rules for creating CI jobs
that are basically a disjunction, and maybe the nesting makes it work as
a conjunction.

So lets get rid of the nesting.