Fix #1382: Make diff_ci.sh work for all MRs
diff_ci.sh attempts to compute the set of files changed by the current
branch (compared against the main branch).
This worked, but not for MRs that are created from a branch with the
name "main", because it would compare the branch against itself.

Luckily, GitLab CI provides a pre-defined set of variables
(https://docs.gitlab.com/ee/ci/variables/predefined_variables.html#predefined-variables-for-merge-request-pipelines),
and for MRs there is the variable CI_MERGE_REQUEST_DIFF_BASE_SHA
that tells us what we need to compare against.

For any other cases where the script may be executed
(pipelines outside of MR contexts, local execution)
we keep the existing behavior of comparing against origin/main...,
i.e., against the merge-base of origin/main and HEAD.
But while we are at it, we can avoid hard-coding the branch name "main"
in some cases.