Fix diff computation of CI script for "merged results" pipelines
In this case, the computed diff would be too large because it would
potentially include changes to the main branch.
When updating the script, I mistakenly assumed that it does not matter
than in a "merged results" pipeline the current HEAD is a merge commit
of the MR branch with the target branch.
This does indeed not matter when computing a "merge base",
which we did before.
But it does matter if we just compute a diff between revisions,
and the updated script does just that.
So we have to take care of this fact and always use exactly the tip of
the MR for MR pipelines, whether "merged results" is active or not.