Merge set files SoftwarSystems-Other-* into single SoftwareSystems-Other
This commit removes the property from the set file names
of two set files where this is problematic.

The set file names with the property in the name
are historically grown. In the past,
the C files encoded the properties directly in the name
and the set files selected for something like `arrays/*unreach-call*.i`.
But since the introduction of the YAML task-definition format,
these file names do not reflect the properties of the included tasks
anymore, so the properties should be removed from the set file name.

This also creates a problem when the set files are used with BenchExec
in other contexts than SV-COMP:
When two set files with intersecting file patterns
are used for a single property,
this leads to issues because tasks are included twice.

One such issue occurs now in Test-Comp.
Test-Comp uses SoftwareSystems-Other-* with property coverage-branches,
example:
https://gitlab.com/sosy-lab/test-comp/bench-defs/-/blob/0a6df3fd1d473c458cdc4cdf5ba02b54ad21831f/benchmark-defs/cetfuzz.xml#L60-65

This includes tasks from busybox-1.22.0 and openbsd-6.2 twice.