[nla-digbench] move UB nla-digbench*/freire[12] tasks to todo/
These tasks contain arithmetic on type `double` and cast that to `int`.

The unconstrained (i.e. non-valuebound) ones have undefined behaviour,
which for both properties (no-overflow and unreach-call) isn't allowed.

As I'm not clear what the intention here is, I do not have a fix to
apply. Therefore the tasks are moved to todo/.