busybox: relabel cut-3
And a fixed version of cut-3 named cut-4.  The fix removes the body of
bb_verror_msg whose task is to print things and where was the invalid
dereference. For the reachability task, the removed code is not
important as the program is aborted after the call.

Other properties than valid-memsafety were removed from cut-3
including coverage-branches etc. (coverage properties are
in cut-4 which should be enough).

See #1403.