Fix an invalid dereferece in an LDV file
The program allocates 1736 bytes to pointer
`si4713_subdev_core_ops_group2` on line 6122 and passes this memory to
the call of `si4713_g_ext_ctrls` on line 6315. There the memory is
casted to a bigger struct `si4713_device` on line 5549 and the allocated
memory is then accessed with offset 2596 on line 5672.

Fix the allocation size to the actual size of the bigger struct.

Partially fixes #1406.