Remove expected verdict of tasks that depend on x87 float semantics
The counterexample that was the reason for labeling these tasks as FALSE
only happens if non-standards-compliant x87 floating-point semantics
(with incorrect keeping of intermediate results with 80 bits of precision)
are used. If correct IEEE754-compliant rounding of intermediate results
is used, the counterexample does not exist.

For an extensive discussion please cf. #1426.